Move fragment controller into their own package
