+++ app/src/main/java/org/cpen321/discovr/fragment/EventsSubscribedFragment.java
@@ -1,4 +1,4 @@
-package org.cpen321.discovr;
+package org.cpen321.discovr.fragment;
 
 
 import android.graphics.drawable.Drawable;
@@ -18,7 +18,10 @@ import android.widget.FrameLayout;
 import android.widget.LinearLayout;
 import android.widget.ScrollView;
 
+import org.cpen321.discovr.R;
+import org.cpen321.discovr.SQLiteDBHandler;
 import org.cpen321.discovr.model.EventInfo;
+import org.cpen321.discovr.fragment.partial.EventPartialFragment;
 
 import java.util.List;
 
@@ -66,7 +69,7 @@ public class EventsSubscribedFragment extends Fragment {
             button.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
-                    SingleEventFragment fragment = new SingleEventFragment();
+                    EventPartialFragment fragment = new EventPartialFragment();
                     FragmentManager fm = getActivity().getSupportFragmentManager();
                     Fragment currentFrag = fm.findFragmentById(R.id.fragment_container);
                     Log.d("backstack", "From Subscribed Events: currFragment = " + currentFrag);
