+++ app/src/main/res/layout/fragment_map_view.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="org.cpen321.discovr.MapViewFragment">
+    tools:context=".fragment.MapViewFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <com.mapbox.mapboxsdk.maps.MapView
