+++ app/src/main/res/layout/fragment_transit.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.cpen321.discovr.TransitFragment"
+    tools:context=".fragment.partial.TransitPartialFragment"
     android:layout_marginTop="400dp"
     android:gravity="top"
     android:weightSum="1"
